Fix issue with the previous path being used for navigation in too many places and refactor the Consent component to a functional component.
