+++ src/components/LoginConsent/ProvisionIdentity/ProvisionIdentityForm/ProvisionIdentityForm.js
@@ -29,7 +29,6 @@ const ProvisionIdentityForm = () => {
   const deeplinkData = useSelector((state) => state.deeplink.data);
   const chainId = useSelector((state) => state.chainMetadata.chainId);
   const chainName = useSelector((state) => state.chainMetadata.chainName);
-  const previousPath = useSelector((state) => state.navigation.previousPath);
   const identityToProvisionField = useSelector((state) => state.provision.identityToProvisionField);
   const initialPrimaryAddress = useSelector((state) => state.provision.primaryAddress);
 
@@ -246,7 +245,7 @@ const ProvisionIdentityForm = () => {
   };
 
   const cancel = () => {
-    dispatch(setNavigationPath(previousPath || SELECT_LOGIN_ID));
+    dispatch(setNavigationPath(SELECT_LOGIN_ID));
   };
 
   return (
